refactor(deployment): move helpers folder
