+++ tasks/deploy-chain-manager-root.ts
@@ -1,8 +1,8 @@
 import { task } from "hardhat/config";
-import { deployRootChainManager } from "./utils/deploy-root-chain-manager";
+import { deployRootChainManager } from "./helpers/deploy-root-chain-manager";
 import { AddressConstants } from "../src/common/constants";
-import { getNetworkEnv } from "./utils/helpers";
-import { verifyContract } from "./utils/verify-contract";
+import { getNetworkEnv } from "./utils";
+import { verifyContract } from "./helpers/verify-contract";
 
 task("deploy:chain-manager:root")
   .setDescription("Deploys the root chain manager")
