+++ tasks/helpers/deploy-child-chain-manager.ts
@@ -1,35 +1,36 @@
+/* eslint-disable import/no-extraneous-dependencies */
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { Signer } from "ethers";
-import { TradeTrustERC721RootTunnel } from "@tradetrust/contracts";
-import { StatusManager } from "./status-manager/status-manager";
+import { TradeTrustERC721ChildTunnel } from "@tradetrust/contracts";
+import { StatusManager } from "../utils/status-manager/status-manager";
 
-type DeployRootChainManagerConstructorParams = {
-  checkPointManager: string;
-  fxRoot: string;
-  rootToken: string;
+type DeployChildChainManagerConstructorParams = {
+  fxChild: string;
+  childToken: string;
 };
 
-export const deployRootChainManager = async ({
+export const deployChildChainManager = async ({
   constructorParams,
   hre,
   deployer,
 }: {
-  constructorParams: DeployRootChainManagerConstructorParams;
+  constructorParams: DeployChildChainManagerConstructorParams;
   hre: HardhatRuntimeEnvironment;
   deployer: Signer;
 }) => {
   const { ethers } = hre;
-  const { checkPointManager, fxRoot, rootToken } = constructorParams;
-  const contractName = "TradeTrustERC721RootTunnel";
+  const { fxChild, childToken } = constructorParams;
+  const contractName = "TradeTrustERC721ChildTunnel";
   const status = StatusManager.create();
 
   try {
     const deployerAddress = await deployer.getAddress();
-    status.add(`Deploying root chain manager contract ${contractName} as deployer ${deployerAddress}...`);
+    status.add(`Deploying child chain manager contract ${contractName} as deployer ${deployerAddress}...`);
+
     const chainManagerFactory = await ethers.getContractFactory(contractName);
     const chainManager = (await chainManagerFactory
       .connect(deployer)
-      .deploy(checkPointManager, fxRoot, rootToken)) as TradeTrustERC721RootTunnel;
+      .deploy(fxChild, childToken)) as TradeTrustERC721ChildTunnel;
 
     const tx = chainManager.deployTransaction;
     status.update(`Pending deployment transaction ${tx.hash}...`);
@@ -39,7 +40,7 @@ export const deployRootChainManager = async ({
 
     return chainManager;
   } catch (err) {
-    status.fail(`Root chain manager deployment failed: ${err}`);
+    status.fail(`Child chain manager deployment failed: ${err}`);
     throw err;
   }
 };
