+++ tasks/helpers/deploy-root-chain-manager.ts
@@ -1,36 +1,35 @@
-/* eslint-disable import/no-extraneous-dependencies */
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { Signer } from "ethers";
-import { TradeTrustERC721ChildTunnel } from "@tradetrust/contracts";
-import { StatusManager } from "./status-manager/status-manager";
+import { TradeTrustERC721RootTunnel } from "@tradetrust/contracts";
+import { StatusManager } from "../utils/status-manager/status-manager";
 
-type DeployChildChainManagerConstructorParams = {
-  fxChild: string;
-  childToken: string;
+type DeployRootChainManagerConstructorParams = {
+  checkPointManager: string;
+  fxRoot: string;
+  rootToken: string;
 };
 
-export const deployChildChainManager = async ({
+export const deployRootChainManager = async ({
   constructorParams,
   hre,
   deployer,
 }: {
-  constructorParams: DeployChildChainManagerConstructorParams;
+  constructorParams: DeployRootChainManagerConstructorParams;
   hre: HardhatRuntimeEnvironment;
   deployer: Signer;
 }) => {
   const { ethers } = hre;
-  const { fxChild, childToken } = constructorParams;
-  const contractName = "TradeTrustERC721ChildTunnel";
+  const { checkPointManager, fxRoot, rootToken } = constructorParams;
+  const contractName = "TradeTrustERC721RootTunnel";
   const status = StatusManager.create();
 
   try {
     const deployerAddress = await deployer.getAddress();
-    status.add(`Deploying child chain manager contract ${contractName} as deployer ${deployerAddress}...`);
-
+    status.add(`Deploying root chain manager contract ${contractName} as deployer ${deployerAddress}...`);
     const chainManagerFactory = await ethers.getContractFactory(contractName);
     const chainManager = (await chainManagerFactory
       .connect(deployer)
-      .deploy(fxChild, childToken)) as TradeTrustERC721ChildTunnel;
+      .deploy(checkPointManager, fxRoot, rootToken)) as TradeTrustERC721RootTunnel;
 
     const tx = chainManager.deployTransaction;
     status.update(`Pending deployment transaction ${tx.hash}...`);
@@ -40,7 +39,7 @@ export const deployChildChainManager = async ({
 
     return chainManager;
   } catch (err) {
-    status.fail(`Child chain manager deployment failed: ${err}`);
+    status.fail(`Root chain manager deployment failed: ${err}`);
     throw err;
   }
 };
